Publish defaultTelemetryFilters
