+++ packages/core/src/types.ts
@@ -251,3 +251,10 @@ export interface GetPluginsOptions {
 export interface GetPluginsOptionsWithCWD extends GetPluginsOptions {
     cwd: string;
 }
+
+export interface Metric {
+    metric: string;
+    type: 'count' | 'size' | 'duration';
+    value: number;
+    tags: string[];
+}
